requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# IBM services
ibm-watsonx-orchestrate  # ADK — also installs orchestrate CLI
ibm-watsonx-ai>=1.0.0
ibm-cloud-sdk-core>=3.18.0

# HTTP client (async Orchestrate REST calls)
httpx>=0.27.0

# Config
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
